requirements.txt
flask
flask-sqlalchemy
pymysql
flask-bcrypt
flask-jwt-extended
jinja2
python-dotenv